Move `download` functions from `native` to builder.rs
This has no logic changes, just a move.
